Fixed more test failures and completed conversions.
