+++ tests/src/Functional/SmsFrameworkPhoneNumberTest.php
@@ -9,6 +9,7 @@ use Drupal\sms\Message\SmsMessageInterface;
  * General phone number verification tests.
  *
  * @group SMS Framework
+ * @group legacy
  */
 class SmsFrameworkPhoneNumberTest extends SmsFrameworkBrowserTestBase {
